cron.d/dsa-buildd: cleanup upload queues
The security upload queue is not cleaned after packages are accepted as
mails are not sent for security reasons. The standard upload queue is not
also not always cleaned as packages can be rejected (new version, out-of
date chroot with Built-Using, etc.)

Automatically cleanup upload queues by removing files older than 60
days. After this time it's probably better to rebuild the package with
an up-to-date chroot.

Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>