Do not install rng-tools on stretch VMs
Newer kernel version, includng the one in stretch, are able to feel the
entropy pool from a hardware random number generator without the help of
userspace. The quality option determine how much entropy is used from
the hardware random number and defaults to the maximum for virtio-rng.

Therefore we don't need rng-tools anymore on stretch VMs.

Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>